refactor: 승자 판단 로직을 Dealer안으로 이동
- 카드를 가지고있는 사람은 Player이다. 특별히 분리하지 않는다면 승패 판단도 카드를 가지고 있는 사람이 가능해야한다고 생각하기떄문에 Dealer로 이동했다.
